CB-15315 Add CDPDatalakeFeatures to CDPDatalakeSync events
Currently only the CDPDatalakeRequested and CDPDatalakeStatusChanged events contain the CDPDatalakeFeatures field that contain information about RAZ status. This info should also be available in the datalake sync events.
