+++ structuredevent-service-legacy/src/test/java/com/sequenceiq/cloudbreak/structuredevent/service/telemetry/converter/StructuredFlowEventToCDPDatahubRequestedConverterTest.java
@@ -20,10 +20,10 @@ public class StructuredFlowEventToCDPDatahubRequestedConverterTest {
         Whitebox.setInternalState(operationDetailsConverter, "appVersion", "version-1234");
         Whitebox.setInternalState(operationDetailsConverter, "clusterRequestProcessingStepMapper", new ClusterRequestProcessingStepMapper());
         Whitebox.setInternalState(underTest, "operationDetailsConverter", operationDetailsConverter);
-        StructuredEventToClusterDetailsConverter clusterDetailsConverter = new StructuredEventToClusterDetailsConverter();
-        Whitebox.setInternalState(clusterDetailsConverter, "clusterShapeConverter", new StructuredEventToClusterShapeConverter());
-        Whitebox.setInternalState(clusterDetailsConverter, "imageDetailsConverter", new StructuredEventToImageDetailsConverter());
-        Whitebox.setInternalState(clusterDetailsConverter, "versionDetailsConverter", new StructuredEventToVersionDetailsConverter());
+        StructuredEventToCDPClusterDetailsConverter clusterDetailsConverter = new StructuredEventToCDPClusterDetailsConverter();
+        Whitebox.setInternalState(clusterDetailsConverter, "clusterShapeConverter", new StructuredEventToCDPClusterShapeConverter());
+        Whitebox.setInternalState(clusterDetailsConverter, "imageDetailsConverter", new StructuredEventToCDPImageDetailsConverter());
+        Whitebox.setInternalState(clusterDetailsConverter, "versionDetailsConverter", new StructuredEventToCDPVersionDetailsConverter());
         Whitebox.setInternalState(underTest, "clusterDetailsConverter", clusterDetailsConverter);
     }
