+++ structuredevent-service-legacy/src/test/java/com/sequenceiq/cloudbreak/structuredevent/service/telemetry/converter/StructuredSyncEventToCDPDatalakeSyncConverterTest.java
@@ -20,13 +20,14 @@ public class StructuredSyncEventToCDPDatalakeSyncConverterTest {
         Whitebox.setInternalState(operationDetailsConverter, "appVersion", "version-1234");
         Whitebox.setInternalState(operationDetailsConverter, "clusterRequestProcessingStepMapper", new ClusterRequestProcessingStepMapper());
         Whitebox.setInternalState(underTest, "operationDetailsConverter", operationDetailsConverter);
-        StructuredEventToClusterDetailsConverter clusterDetailsConverter = new StructuredEventToClusterDetailsConverter();
-        Whitebox.setInternalState(clusterDetailsConverter, "clusterShapeConverter", new StructuredEventToClusterShapeConverter());
-        Whitebox.setInternalState(clusterDetailsConverter, "imageDetailsConverter", new StructuredEventToImageDetailsConverter());
-        Whitebox.setInternalState(clusterDetailsConverter, "versionDetailsConverter", new StructuredEventToVersionDetailsConverter());
+        StructuredEventToCDPClusterDetailsConverter clusterDetailsConverter = new StructuredEventToCDPClusterDetailsConverter();
+        Whitebox.setInternalState(clusterDetailsConverter, "clusterShapeConverter", new StructuredEventToCDPClusterShapeConverter());
+        Whitebox.setInternalState(clusterDetailsConverter, "imageDetailsConverter", new StructuredEventToCDPImageDetailsConverter());
+        Whitebox.setInternalState(clusterDetailsConverter, "versionDetailsConverter", new StructuredEventToCDPVersionDetailsConverter());
         Whitebox.setInternalState(underTest, "clusterDetailsConverter", clusterDetailsConverter);
         Whitebox.setInternalState(underTest, "syncDetailsConverter", new StructuredSyncEventToCDPSyncDetailsConverter());
-        Whitebox.setInternalState(underTest, "statusDetailsConverter", new StructuredEventToStatusDetailsConverter());
+        Whitebox.setInternalState(underTest, "statusDetailsConverter", new StructuredEventToCDPStatusDetailsConverter());
+        Whitebox.setInternalState(underTest, "featuresConverter", new StructuredEventToCDPDatalakeFeaturesConverter());
     }
 
     @Test
@@ -37,6 +38,7 @@ public class StructuredSyncEventToCDPDatalakeSyncConverterTest {
         Assertions.assertNotNull(datalakeSync.getSyncDetails());
         Assertions.assertNotNull(datalakeSync.getClusterDetails());
         Assertions.assertNotNull(datalakeSync.getStatusDetails());
+        Assertions.assertNotNull(datalakeSync.getFeatures());
     }
 
     @Test
@@ -48,5 +50,6 @@ public class StructuredSyncEventToCDPDatalakeSyncConverterTest {
         Assertions.assertNotNull(datalakeSync.getSyncDetails());
         Assertions.assertNotNull(datalakeSync.getClusterDetails());
         Assertions.assertNotNull(datalakeSync.getStatusDetails());
+        Assertions.assertNotNull(datalakeSync.getFeatures());
     }
 }
